backend/requirements-cloud.txt
# Cloud-only Python dependencies for the Moonwalk Brain
# These are the packages needed on Google Cloud Run
# (no macOS-specific packages like pyobjc, pvporcupine, etc.)

google-genai>=1.0.0
python-dotenv>=1.0.0
websockets>=12.0
httpx>=0.27.0
numpy>=1.24.0
